list: ListRemoveNode() was not updating 'prev'
This was causing the first node to point to the node after the
deleted node, rather than the node before the deleted node pointing
to the node after the deleted node. In short, it was a mess, which
really shortened the lists quickly and caused nodes to be lost once
deletion took place of a node that was not the first in the list.

Signed-off-by: Joel Nider <joel@nider.org>