Complete the move to C++
Well, not completely in C++, but the main components are there.
The rest is just bug fixes.  It looks like the 'board' will completely
disappear, having been taken over by the 'game'. All of the game logic
is now in 'game'. It is now set up to handle different UI implementations
by implementing different subclasses of CGameUI. Some config file or
command line option would be needed to select a different UI.
The linked list implementation can now be replaced by something more
standard like stl. Some of the SDL UI was not properly implemented, and
will have to be redone.

Signed-off-by: Joel Nider <joel@nider.org>